requirements.txt
trimesh
scikit-learn
opencv-python 
matplotlib
imageio
diffusers
einops
transformers
gradio
omegaconf
triton
accelerate
objaverse
basicsr==1.4.2
numpy==1.26.4
pytorch-lightning
open-clip-torch
xatlas